Create task6.m
This task is to model a Solar Heating System and estimate the parameters of the system using Least Squares Methods. Instead of the system is non-linear by nature, it is linear on the parameters. 
The data were collected using the site http://arohatgi.info/WebPlotDigitizer/

References on TOOLS FOR SEMIPHYSICAL MODELLING.pdf